+++ src/Utilities.cpp
@@ -42,7 +42,7 @@ bool benchmark_script;
 map<set<Region*>, set<int>*> *intersection_cache;
 int places_after_initial_decomp;
 int places_after_greedy;
-double maxAlphabet;
+int maxAlphabet;
 double avgAlphabet;
 bool greedy_exact;
 bool check_structure;
